Clean up MIDI file naming conventions
- Combine simpleButtonMidiSend/Receive into noiseRingModMidi.ts
- Rename outMidi.ts to outPostMixMidi.ts for clarity
- Rename test file to match
- Update all imports in index.tsx and tests
